requirements.txt
google-generativeai==0.3.2
PyPDF2==3.0.1
faiss-cpu==1.7.4
numpy==1.24.3
torch==2.1.0
transformers==4.30.2
huggingface-hub==0.16.4
accelerate==0.20.3
sentence-transformers==2.2.2
streamlit==1.32.0
pdfplumber==0.10.3 